update: useHighChartsにロジックを切り出し
