+++ src/features/prefecture-population/components/population-graph/PopulationGraph.module.css
@@ -1,0 +1,4 @@
+.title {
+  margin-left: 10px;
+  font-size: 20px;
+}
